polymorphism with unit testing

+++ CoreJava/src/com/training/JewelLoan.java
@@ -4,7 +4,7 @@ public class JewelLoan extends Loan {
 
 	//private long gram;
 	
-	public long interestCalculate(long gram){
+	public double interestCalculate(double gram){
 		
 		if(gram < 100){
 			return  10;
